requirements.txt
# Core dependencies for camera and AI
# run "pip install -r requirements.txt" in terminal to update packages
numpy<2
requests
Pillow
torch
transformers
einops 
pulp 
scipy
shapely
matplotlib
timm
